Moved DSK on air action from DownstreamKey group to MixEffect group according to isse #56
