+++ src/functions/downstreamKeyer/feedbackId.ts
@@ -1,4 +1,3 @@
 export enum FeedbackId {
-	DskOnAir = 'dskOnAir',
 	DskSourceFill = 'dskSourceFill',
 }
