+++ src/functions/mixEffect/presets.ts
@@ -576,5 +576,48 @@ export function create(model: GoStreamModel): CompanionPresetDefinitions {
 			},
 		],
 	}
+
+	for (const key of Keys) {
+		if (key.label != 'BKGD' && key.label !== 'Key') {
+			presets[`keys_Next_Air_${key.id}`] = {
+				category: 'Keys On Air',
+				name: `Toggle upstream KEY ${key.label} OnAir`,
+				type: 'button',
+				style: {
+					text: `${key.label}`,
+					size: ptzSize,
+					color: combineRgb(255, 255, 255),
+					bgcolor: combineRgb(0, 0, 0),
+				},
+				feedbacks: [
+					{
+						feedbackId: FeedbackId.DskOnAir,
+						options: {
+							KeyOnAir: 1,
+							DSKOnAir: 1,
+						},
+						style: {
+							bgcolor: combineRgb(255, 255, 0),
+							color: combineRgb(0, 0, 0),
+						},
+					},
+				],
+				steps: [
+					{
+						down: [
+							{
+								actionId: ActionId.DskOnAir,
+								options: {
+									KeyOnAir: 2,
+									DSKOnAir: 2,
+								},
+							},
+						],
+						up: [],
+					},
+				],
+			}
+		}
+	}
 	return presets
 }
